console: header rename according to the platform conventions
