+++ platform/host/export/platform/console.hpp
