+++ platform/stm32f4xx/templates/console.hpp.in
