+++ platform/stm32f4xx/templates/console_stub.hpp.in
