+++ platform/stm32l1xx/templates/console.hpp.in
