+++ platform/stm32l1xx/templates/console_stub.hpp.in
